Use grid-access to prevent redrawing too often
It seems previously it was possible to draw frames as often as the
GPU would allow between grid updates. Now the thread will keep the
lock during its sleep period which makes the draw loop compete
for it more, and not attempt to draw more often than the grid is
updated.

Ideally, there would be a variable that keeps track of wheter a redraw
is necessary or not.